+++ src/main/java/com/mraof/minestuck/player/ClientPlayerData.java
@@ -6,6 +6,7 @@ import com.mraof.minestuck.inventory.captchalogue.CaptchaDeckHandler;
 import com.mraof.minestuck.inventory.captchalogue.Modus;
 import com.mraof.minestuck.alchemy.GristSet;
 import com.mraof.minestuck.network.ColorSelectPacket;
+import com.mraof.minestuck.network.data.EditmodeCacheLimitPacket;
 import com.mraof.minestuck.network.MSPacketHandler;
 import com.mraof.minestuck.network.RGBColorSelectPacket;
 import com.mraof.minestuck.network.data.*;
@@ -34,6 +35,7 @@ public final class ClientPlayerData
 	private static int consortReputation;
 	private static GristSet playerGrist;
 	private static GristSet targetGrist;
+	private static long targetCacheLimit;
 	private static int playerColor;
 	private static boolean displaySelectionGui;
 	private static boolean dataCheckerAccess;
@@ -81,14 +83,17 @@ public final class ClientPlayerData
 		return consortReputation;
 	}
 	
-	public static GristSet getGristCache(CacheSource cacheSource)
+	public static ClientCache getGristCache(CacheSource cacheSource)
 	{
 		return switch(cacheSource)
 		{
-			case PLAYER -> ClientPlayerData.playerGrist;
-			case EDITMODE -> ClientPlayerData.targetGrist;
+			case PLAYER -> new ClientCache(ClientPlayerData.playerGrist, Echeladder.getGristCapacity(ClientPlayerData.getRung()));
+			case EDITMODE -> new ClientCache(ClientPlayerData.targetGrist, targetCacheLimit);
 		};
 	}
+	
+	public record ClientCache(GristSet set, long limit)
+	{}
 	
 	public enum CacheSource
 	{
@@ -164,6 +169,11 @@ public final class ClientPlayerData
 		if(packet.isEditmode)
 			targetGrist = packet.gristCache;
 		else playerGrist = packet.gristCache;
+	}
+	
+	public static void handleDataPacket(EditmodeCacheLimitPacket packet)
+	{
+		targetCacheLimit = packet.limit();
 	}
 	
 	public static void handleDataPacket(ColorDataPacket packet)